+++ src/mako/error/handlers/web/ProductionHandler.php
@@ -144,7 +144,14 @@ class ProductionHandler implements HandlerInterface
 			}
 		}
 
-		return $this->view->clearAutoAssignVariables()->render('mako-error::' . $view);
+		try
+		{
+			return $this->view->render('mako-error::' . $view);
+		}
+		catch(Throwable $e)
+		{
+			return $this->view->clearAutoAssignVariables()->render('mako-error::' . $view);
+		}
 	}
 
 	/**